SHM System/requirements.txt
# Secure IoT SHM System - Dependencies
# ======================================

# MQTT Client
paho-mqtt==1.6.1

# FastAPI + ASGI server
fastapi==0.111.0
uvicorn[standard]==0.29.0

# JWT Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data validation
pydantic==2.7.1

# TLS Certificate Generation
cryptography==42.0.5

# ML Anomaly Detection (optional)
scikit-learn==1.4.2
numpy==1.26.4

# CORS
python-multipart==0.0.9
